requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.3
bcrypt==4.1.2